Runnin with the Devil official stems/audio quality upgrade for reprocessed moggs
This updates the audio for Van Halen - Runnin' with the Devil to base it on the official RB4 stems (with the GHVH crowd tracking retained, and adjusted to better account for the different stems compared to the previous mogg) and reencodes the moggs for songs that needed to have their moggs reprocessed (decensors, songs with crowd tracking added, etc) to have a higher quality to be closer to the official quality. Also most of the decensors have had the diy portions for vocals redone in UVR with the UVR-MDX-NET Inst HQ 3 model.